+++ __init__.py
@@ -3,5 +3,3 @@ from .nodes.nodes import NODE_CLASS_MAPPINGS, NODE_DISPLAY_NAME_MAPPINGS
 WEB_DIRECTORY = "./js"
 
 __all__ = ['NODE_CLASS_MAPPINGS','NODE_DISPLAY_NAME_MAPPINGS','WEB_DIRECTORY']
-
-print("\033[34mComfyUI_simpleButcher Node Suite: \033[92mLoaded\033[0m")