+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # nightly
+# v1.0.5
+- Add node `Simple Load Images From Dir` to batch images
+- Add `fixed` variant to `Simple Load Line From Text File`
 - `Simple Image Saver` - now when saving images, optimization is enabled to reduce file size.
 # v1.0.4
 - update `Simple Load Image With Metadata` and `Simple Image Saver (as Forge)`